requirements.txt
Tested on: Ubuntu 18
Note: Runs a local server on http://127.0.0.1:5000 (flask), port must be free

Need to have installed:
Python v.3+
And these Python libraries:
fusepy
flask
pywebview
pydal

Install commands for pip (note: pip3 for Python3):
pip3 install fusepy
pip3 install flask
pip3 install pywebview
pip3 install pydal
